src/aiserver/LLM_Facial_Memory_System/requirements.txt
numpy
opencv-python
insightface
python-dotenv
openai
watchdog
onnxruntime
pillow
